ComplexFuncBench/pyproject.toml
[project]
name = "complexfuncbench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.62.0",
    "flagembedding>=1.3.5",
    "mistralai>=1.9.3",
    "numpy>=2.3.2",
    "openai>=1.99.9",
    "python-dotenv>=1.1.1",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "zhipuai>=2.1.5.20250801",
]
